Use the provider network for router id
The router id for the dragent must be on the provider network. The
charm was using the BGP speaker_ip which is only used for related
test services. (i.e. quagga).

This change guarantees the router id uses the provider extra binding
IP.

It also sets the group ownership to neutron for rendered config files.

Closes-Bug: #1782433
Partial-Bug: #1782444
Change-Id: I4761b4c5b80a398e13e4cdf880b773df489f53b9