Add data plane testing for NDR routes
The change includes modifications to bundles to set up components
necessary to spawn instances and provide actual connectivity to them. It
makes the testing more heavy but adds assurances that the data plane
works for the routes advertised by the NDR control plane.

The bundle changes also fix an issue that got in with the original OVN
bundles: manage-neutron-plugin-legacy-mode is set to True and the plugin
is determined to be OVS instead of OVN. The control plane for NDR still
worked but with a non-functioning ML2/ovs config causing OVN DBs to be
empty.

Func-Test-PR: https://github.com/openstack-charmers/zaza-openstack-tests/pull/1076
Change-Id: Ie59b942a0800ce8dd979398f41ed2138472481f1